Testing bedrock/ skyrock, added unique layers
Temporarily added a bedrock and skyrock node (in default namespace for now on my local machine as it is in the default mod ... sorry ...). Introduced a layer table containing and a loop to populate the layer settings based on a single random seed and with some noise settings hard coded for now.